scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "tqdm>=4.67.1",
]
